+++ algorithm.py
@@ -1,914 +1,218 @@
-import shutil
-from fastapi import HTTPException
 import pulp
-from collections import defaultdict
-import json
-
+import sqlite3
 def load_days():
     return ['Monday', 'Tuesday', 'Wednesday']
 
 def load_periods():
     mwf_periods = {
-    1: {'start_time': '8:00AM', 'duration': 50, 'end_time': '8:50AM'},
-    2: {'start_time': '9:05AM', 'duration': 50, 'end_time': '9:55AM'},
-    3: {'start_time': '10:10AM', 'duration': 50, 'end_time': '11:00AM'},
-    4: {'start_time': '11:15AM', 'duration': 50, 'end_time': '12:05PM'},
-    5: {'start_time': '12:20PM', 'duration': 50, 'end_time': '1:10PM'},
-    6: {'start_time': '1:25PM', 'duration': 50, 'end_time': '2:15PM'},
-    7: {'start_time': '2:30PM', 'duration': 50, 'end_time': '3:20PM'},
-    8: {'start_time': '3:35PM', 'duration': 50, 'end_time': '4:25PM'},
+        1: {'start_time': '8:00AM', 'duration': 50},
+        2: {'start_time': '9:05AM', 'duration': 50},
+        3: {'start_time': '10:10AM', 'duration': 50},
+        4: {'start_time': '11:15AM', 'duration': 50},
+        5: {'start_time': '12:20PM', 'duration': 50},
+        6: {'start_time': '1:25PM', 'duration': 50},
+        7: {'start_time': '2:30PM', 'duration': 50},
+        8: {'start_time': '3:35PM', 'duration': 50},
 }
 
 # TTH periods (75 minutes each)
     tth_periods = {
-        1: {'start_time': '8:00AM', 'duration': 75, 'end_time': '9:15AM'},
-        2: {'start_time': '9:30AM', 'duration': 75, 'end_time': '10:45AM'},
-        3: {'start_time': '11:00AM', 'duration': 75, 'end_time': '12:15PM'},
-        4: {'start_time': '12:30PM', 'duration': 75, 'end_time': '1:45PM'},
-        5: {'start_time': '2:00PM', 'duration': 75, 'end_time': '3:15PM'},
-        6: {'start_time': '3:30PM', 'duration': 75, 'end_time': '4:45PM'},
-        7: {'start_time': '5:00PM', 'duration': 75, 'end_time': '6:15PM'}
+        1: {'start_time': '8:00AM', 'duration': 75},
+        2: {'start_time': '9:30AM', 'duration': 75},
+        3: {'start_time': '11:00AM', 'duration': 75},
+        4: {'start_time': '12:30PM', 'duration': 75},
+        5: {'start_time': '2:00PM', 'duration': 75},
+        6: {'start_time': '3:30PM', 'duration': 75},
+        7: {'start_time': '5:00PM', 'duration': 75}
     }
 
-    # MW periods (75 minutes each)
+# MW periods (75 minutes each)
     mw_periods = {
-        1: {'start_time': '8:00AM', 'duration': 75, 'end_time': '9:15AM'},
-        2: {'start_time': '9:05AM', 'duration': 75, 'end_time': '10:20AM'},
-        3: {'start_time': '10:10AM', 'duration': 75, 'end_time': '11:25AM'},
-        4: {'start_time': '11:15AM', 'duration': 75, 'end_time': '12:30PM'},
-        5: {'start_time': '12:20PM', 'duration': 75, 'end_time': '1:35PM'},
-        6: {'start_time': '1:25PM', 'duration': 75, 'end_time': '2:40PM'},
-        7: {'start_time': '2:30PM', 'duration': 75, 'end_time': '3:45PM'},
-        8: {'start_time': '3:35PM', 'duration': 75, 'end_time': '4:50PM'},
-    }
+    1: {'start_time': '8:00AM', 'duration': 75},
+    3: {'start_time': '10:10AM', 'duration': 75},
+    5: {'start_time': '12:20PM', 'duration': 75},
+    7: {'start_time': '2:30PM', 'duration': 75},
+}
     return mwf_periods, tth_periods, mw_periods
 
 
 def load_meeting_patterns(mwf_periods, tth_periods, mw_periods):
     meeting_patterns = {
-        'MWF': {'days': ['Monday'], 'periods': mwf_periods},
-        'TTH': {'days': ['Tuesday'], 'periods': tth_periods},
-        'MW': {'days': ['Monday'], 'periods': mw_periods},
-        }  
+    'MWF': {'days': ['Monday', 'Wednesday', "Friday"], 'periods': mwf_periods},
+    'TTH': {'days': ['Tuesday', 'Thursday'], 'periods': tth_periods},
+    'MW': {'days': ['Monday', "Wednesday"], 'periods': mw_periods},
+    #'WF': {'days': ['Wednesday', 'Friday'], 'periods': wf_periods}
+}
     return meeting_patterns
 
 
-def backup_data():
-    FILE_PATH = "professors.json"
-    BACKUP_PATH = "professors_backup.json"
+
+def create_professors_data():
+    # Fetch qualified courses
+    connection = sqlite3.connect("database.db")
+    connection.row_factory = sqlite3.Row
+    cursor = connection.cursor()
     try:
-        shutil.copy(FILE_PATH, BACKUP_PATH)
-    except FileNotFoundError:
-        with open(BACKUP_PATH, "w") as backup_file:
-            json.dump({}, backup_file, indent=4)
+        cursor.execute("SELECT * FROM QualifiedCourses")
+        qcourses = [dict(row) for row in cursor.fetchall()]
+    except sqlite3.Error as error:
+        print(f"Error occurred while fetching qualified courses: {error}")
+        qcourses = []
+    finally:
+        connection.close()
 
-def restore_backup():
-    FILE_PATH = "professors.json"
-    BACKUP_PATH = "professors_backup.json"
+    # Fetch distinct professors from Availability
+    connection = sqlite3.connect("database.db")
+    connection.row_factory = sqlite3.Row
+    cursor = connection.cursor()
     try:
-        shutil.copy(BACKUP_PATH, FILE_PATH)
-        print("Rollback successful. Changes have been undone.")
-    except FileNotFoundError:
-        raise HTTPException(status_code=500, detail="Backup file not found. Cannot restore.")
+        cursor.execute("SELECT DISTINCT Prof FROM Availability")
+        profs = [dict(row) for row in cursor.fetchall()]
+    except sqlite3.Error as error:
+        print(f"Error occurred while fetching professors: {error}")
+        profs = []
+    finally:
+        connection.close()
 
-def get_professordata():
-    file_path = "professors.json"
+    # Convert list of dicts to a list of professor names
+    new_profs = [prof['Prof'] for prof in profs]
 
-    with open(file_path) as file:
-        data = json.load(file)
-    return data
-
-def add_professor(name, qualified_courses, availability, max_classes):
-    backup_data()
-    file_path = "professors.json"
-
-    try:
-        with open(file_path, 'r') as file:
-            data = json.load(file)
-    except FileNotFoundError:
-        data = {}
-
-    if name in data:
-        print(f"Professor {name} already exists in the dataset.")
-        return data
-    data[name] = {
-        "qualified_courses": qualified_courses,
-        "availability": availability,
-        "max_classes": max_classes
-    }
-    with open(file_path, 'w') as file:
-        json.dump(data, file, indent=4)
-    return data
-
-def update_professor(name, qualified_courses = None, availability = None, max_classes = None):
-    backup_data()
-    file_path = "professors.json"
-    try: 
-        with open(file_path, 'r') as file:
-            data = json.load(file)
-    except FileNotFoundError:
-        raise HTTPException(status_code=404, detail="No data file found. Cannot update.")
-    if name not in data:
-        raise HTTPException(status_code=404, detail=f"Professor {name} not found.")
-    professor = data[name]
-    if qualified_courses is not None:
-        professor["qualified_courses"] = qualified_courses
-    if availability is not None:
-        professor["availability"] = availability
-    if max_classes is not None:
-        professor["max_classes"] = max_classes
-    with open(file_path, 'w') as file:
-        json.dump(data, file, indent=4)
-    return data
-
-def delete_professor(name):
-    backup_data()
-    file_path = "professors.json"
-    try: 
-        with open(file_path, 'r') as file:
-            data = json.load(file)
-    except FileNotFoundError:
-        raise HTTPException(status_code=404, detail="No data file found. Cannot delete.")
-    if name not in data:
-        raise HTTPException(status_code=404, detail=f"Professor {name} not found.")  
-    del data[name]
-    with open(file_path, 'w') as file:
-        json.dump(data, file, indent=4)    
-    return data 
-
-def backup_courses():
-    FILE_PATH = "course.json"
-    BACKUP_PATH = "courses_backup.json"
-    try:
-        shutil.copy(FILE_PATH, BACKUP_PATH)
-    except FileNotFoundError:
-        with open(BACKUP_PATH, "w") as backup_file:
-            json.dump({}, backup_file, indent=4)
-
-def restore_courses():
-    FILE_PATH = "courses.json"
-    BACKUP_PATH = "courses_backup.json"
-    try:
-        shutil.copy(BACKUP_PATH, FILE_PATH)
-        print("Rollback successful. Changes have been undone.")
-    except FileNotFoundError:
-        raise HTTPException(status_code=500, detail="Backup file not found. Cannot restore.")
-
-def get_coursedata():
-    file_path = "courses.json"
-    try:
-        with open(file_path, 'r') as file:
-            data = json.load(file)    
-    except FileNotFoundError: 
-        raise HTTPException(status_code=404, detail = "No data file found.") 
-    return data
-
-def add_courses(title, sections):
-    #backup_courses()
-    file_path = "courses.json"
-    try:
-        with open(file_path, 'r') as file:
-            data = json.load(file)
-    except FileNotFoundError:
-        data = {}
-
-    if title in data:
-        print(f"Course {title} already exists in the dataset.")
-        return data
-    data[title] = {
-        "sections": sections
-    }
-    with open(file_path, 'w') as file:
-        json.dump(data, file, indent=4)
-    return data
-
-def update_course(title, sections):
-    #backup_courses()
-    file_path = "courses.json"
-    try: 
-        with open(file_path, 'r') as file:
-            data = json.load(file)
-    except FileNotFoundError:
-        raise HTTPException(status_code=404, detail="No data file found. Cannot update.")
-    if title not in data:
-        raise HTTPException(status_code=404, detail=f"Professor {name} not found.")
-    course = data[title]
-    if sections is not None:
-        course["sections"] = sections
-
-    with open(file_path, 'w') as file:
-        json.dump(data, file, indent=4)
-    return data
-
-def delete_course(title):
-    #backup_courses()
-    file_path = "courses.json"
-    try: 
-        with open(file_path, 'r') as file:
-            data = json.load(file)
-    except FileNotFoundError:
-        raise HTTPException(status_code=404, detail="No data file found. Cannot delete.")
-    if title not in data:
-        raise HTTPException(status_code=404, detail=f"Course {title} not found.")  
-    del data[title]
-    with open(file_path, 'w') as file:
-        json.dump(data, file, indent=4)    
-    return data 
-
-def get_roomdata():
-    file_path = "rooms.json"
-    try:
-        with open(file_path, 'r') as file:
-            data = json.load(file)    
-    except FileNotFoundError: 
-        raise HTTPException(status_code=404, detail = "No data file found.") 
-    return data
-
-def add_room(name, capacity):
-    file_path = "rooms.json"
-    try:
-        with open(file_path, 'r') as file:
-            data = json.load(file)
-    except FileNotFoundError:
-        data = {}
-
-    if name in data:
-        print(f"Room {name} already exists in the dataset.")
-        return data
-    data[name] = {
-        "capacity": capacity
-    }
-    with open(file_path, 'w') as file:
-        json.dump(data, file, indent=4)
-    return data    
-
-def update_room(name, capacity):
-    file_path = "rooms.json"
-    try: 
-        with open(file_path, 'r') as file:
-            data = json.load(file)
-    except FileNotFoundError:
-        raise HTTPException(status_code=404, detail="No data file found. Cannot update.")
-    if name not in data:
-        raise HTTPException(status_code=404, detail=f"Professor {name} not found.")
-    room = data[name]
-    room["capacity"] = capacity
-    with open(file_path, 'w') as file:
-        json.dump(data, file, indent=4)
-    return data
-
-def delete_room(name):
-    #backup_courses()
-    file_path = "rooms.json"
-    try: 
-        with open(file_path, 'r') as file:
-            data = json.load(file)
-    except FileNotFoundError:
-        raise HTTPException(status_code=404, detail="No data file found. Cannot delete.")
-    if name not in data:
-        raise HTTPException(status_code=404, detail=f"Course {name} not found.")  
-    del data[name]
-    with open(file_path, 'w') as file:
-        json.dump(data, file, indent=4)    
-    return data 
-
-def load_professors():
-    time_slots = []
-    mwf_periods, tth_periods, mw_periods = load_periods()
-    meeting_patterns = load_meeting_patterns(mwf_periods, tth_periods, mw_periods)
-    for mp in meeting_patterns.values():
-        for day in mp['days']:
-            for period in mp['periods'].keys():
-                time_slots.append((day, period))
-    professors = {
-    'Montek Singh': {
-        'qualified_courses': ['COMP541', 'COMP572'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Tessa Joseph-Nicholas': {
-        'qualified_courses': ['COMP126', "COMP380", "COMP380H"],
-        'availability': time_slots,
-        'max_classes': 3
-    },
-    'Ketan Mayer-Patel': {
-        'qualified_courses': ['COMP301', 'COMP426'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Prairie Goodwin': {
-        'qualified_courses': ['COMP301'],
-        'availability': time_slots,
-        'max_classes': 2
-    },
-    'Sayeed Ghani': {
-        'qualified_courses': ['COMP210'],
-        'availability': time_slots,
-        'max_classes': 2
-    },
-    'P.S. Thiagarajan': {
-        'qualified_courses': ['COMP455'],
-        'availability': time_slots,
-        'max_classes': 2
-    },
-    'Jasleen Kaur': {
-        'qualified_courses': ['COMP431'],
-        'availability': time_slots,
-        'max_classes': 0
-    },
-    'Saba Eskandarian': {
-        'qualified_courses': ['COMP537', 'COMP455', 'COMP435', 'COMP590', 'COMP790'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Ron Alterovitz': {
-        'qualified_courses': ['COMP581', 'COMP781', 'COMP782'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Cynthia Sturton': {
-        'qualified_courses': ['COMP435'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Marc Niethammer': {
-        'qualified_courses': ['COMP775'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Izzi Hinks': {
-        'qualified_courses': ['COMP110'],
-        'availability': time_slots,
-        'max_classes': 2
-    },
-    'Samarjit Chakraborty': {
-        'qualified_courses': ['COMP545', 'COMP790-148'],
-        'availability': time_slots,
-        'max_classes': 2
-    },
-    'Donald Porter': {
-        'qualified_courses': ['COMP530'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    # Note: No longer at UNC
-    'John Majikes': {
-        'qualified_courses': ['COMP421', 'COMP550', 'COMP116'],
-        'availability': time_slots,
-        'max_classes': 2
-    },
-    # Note: Alyssa not on the given spreadsheet
-    'Alyssa Byrnes': {
-        'qualified_courses': ['COMP110', 'COMP116', 'COMP210', 'COMP283'],
-        'availability': time_slots,
-        'max_classes': 2
-    },
-    'Gedas Bertasius': {
-        # Note, cross listed 590 and 790. Should get rid of 790 version
-        'qualified_courses': ['COMP790-170'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Roni Sengupta': {
-        'qualified_courses': ['COMP590-177'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Kevin Sun': {
-        'qualified_courses': ['COMP283', 'COMP455', 'COMP550'],
-        'availability': time_slots,
-        'max_classes': 2
-    },
-    'Cece McMahon': {
-        'qualified_courses': ['COMP311', 'COMP541'],
-        'availability': time_slots,
-        'max_classes': 2
-    },
-    'Shahriar Nirjon': {
-        'qualified_courses': ['COMP433'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Jack Snoeyink': {
-        'qualified_courses': ['COMP283', 'DATA140'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Brent Munsell': {
-        'qualified_courses': ['COMP211', 'COMP590', 'COMP530', 'COMP311', 'COMP116'],
-        'availability': time_slots,
-        'max_classes': 2
-    },
-    'James Anderson': {
-        'qualified_courses': ['COMP737', 'COMP750'],
-        'availability': time_slots,
-        'max_classes': 2
-    },
-    'Danielle Szafir': {
-        # Note, he says he prefers visualization design, but is happy to teach something else
-        # How do we specify what his qualified courses are?
-        'qualified_courses': ['COMP790-172'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Daniel Szafir': {
-        'qualified_courses': ['COMP581'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    # Note: The course he specified was Safe Autonomy, but the course he teaches is Formal Methods
-    'Parasara Sridhar Duggirala': {
-        'qualified_courses': ['COMP089','COMP790-144'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Praneeth Chakravarthula': {
-        'qualified_courses': [ 'COMP790-175'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Ben Berg': {
-        'qualified_courses': ['COMP790-178'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Shashank Srivastava': {
-        'qualified_courses': ['COMP664'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Snigdha Chaturvedi': {
-        'qualified_courses': ['COMP790-158'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Huaxiu Yao': {
-        'qualified_courses': ['COMP790-183'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Andrew Kwong': {
-        'qualified_courses': ['COMP790-185'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Mike Reed': {
-        'qualified_courses': ['COMP475'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Paul Stotts': {
-        'qualified_courses': ['COMP590-59', 'COMP523'],
-        'availability': time_slots,
-        'max_classes': 3
-    },
-    'Prasun Dewan': {
-        'qualified_courses': ['COMP524'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Jorge Silva': {
-        'qualified_courses': ['COMP562'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Kris Jordan': {
-        'qualified_courses': ['COMP590-140'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'Junier Oliva': {
-        'qualified_courses': ['COMP755'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    'TBD': {
-        'qualified_courses': ['COMP421'],
-        'availability': time_slots,
-        'max_classes': 1
-    },
-    
-}
-    return professors
-
-def load_courses():
-    courses = {
-    'COMP110': {
-        'title': 'COMP110 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 300},
-            {'section_number': 2, 'seat_capacity': 300},
-            {'section_number': 3, 'seat_capacity': 150},
-            {'section_number': 4, 'seat_capacity': 150}
-        ]
-    },
-    'COMP116': {
-        'title': 'COMP116 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': None}
-        ]
-    },
-    'COMP126': {
-        'title': 'COMP126 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 120}
-        ]
-    },
-    'COMP210': {
-        'title': 'COMP210 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 210},
-            {'section_number': 2, 'seat_capacity': 210}
-        ]
-    },
-    'COMP211': {
-        'title': 'COMP211 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 200},
-            {'section_number': 2, 'seat_capacity': 200}
-        ]
-    },
-    'COMP227': {
-        'title': 'COMP227 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 30}
-        ]
-    },
-    'COMP283': {
-        'title': 'COMP283/283H Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 204}
-            # Combined 283H section and increased the seat capacity by 24 from 180 to 204
-        ]
-    },
-    'COMP301': {
-        'title': 'COMP301 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 200},
-            {'section_number': 2, 'seat_capacity': 200}
-        ]
-    },
-    'COMP311': {
-        'title': 'COMP311 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 225}
-        ]
-    },
-    'COMP380': {
-        'title': 'COMP380 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 60}
-        ]
-    },
-    'COMP380H': {
-        'title': 'COMP380H Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 24}
-        ]
-    },
-    'COMP421': {
-        'title': 'COMP421 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 200}  # Assuming maximum capacity
-        ]
-    },
-    'COMP426': {
-        'title': 'COMP426 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 200}  # Assuming maximum capacity
-        ]
-    },
-    'COMP431': {
-        'title': 'COMP431 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 60}
-        ]
-    },
-    'COMP433': {
-        'title': 'COMP433 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 75}
-        ]
-    },
-    'COMP435': {
-        'title': 'COMP435 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 70}
-        ]
-    },
-    'COMP455': {
-        'title': 'COMP455 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 125},
-            {'section_number': 2, 'seat_capacity': 125}
-        ]
-    },
-    'COMP475': {
-        'title': 'COMP475 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 60}
-        ]
-    },
-    'COMP488': {
-        'title': 'COMP488 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': None}
-        ]
-    },
-    'COMP520': {
-        'title': 'COMP520 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': None}
-        ]
-    },
-    'COMP523': {
-        'title': 'COMP523 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 60}
-        ]
-    },
-    'COMP524': {
-        'title': 'COMP524 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 60}
-        ]
-    },
-    'COMP530': {
-        'title': 'COMP530 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 75}
-        ]
-    },
-    'COMP533': {
-        'title': 'COMP533 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 60}
-        ]
-    },
-    'COMP537': {
-        'title': 'COMP537 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 90}
-        ]
-    },
-    'COMP541': {
-        'title': 'COMP541 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 45},
-            {'section_number': 2, 'seat_capacity': 45}
-        ]
-    },
-    'COMP545': {
-        'title': 'COMP545 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 25}
-        ]
-    },
-    'COMP550': {
-        'title': 'COMP550 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 125},
-            {'section_number': 2, 'seat_capacity': 125}
-        ]
-    },
-    'COMP560': {
-        'title': 'COMP560 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 60}
-        ]
-    },
-    'COMP562': {
-        'title': 'COMP562 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 75}
-        ]
-    },
-    'COMP581': {
-        'title': 'COMP581 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 60}
-        ]
-    },
-    # 'COMP590': {
-    #     'title': 'COMP590 Class',
-    #     'sections': [
-    #         {'section_number': 59, 'seat_capacity': 90},
-    #         {'section_number': 139, 'seat_capacity': 10},
-    #         {'section_number': 140, 'seat_capacity': 280},
-    #         {'section_number': 158, 'seat_capacity': 4},
-    #         {'section_number': 170, 'seat_capacity': 30},
-    #         {'section_number': 172, 'seat_capacity': 30},
-    #         {'section_number': 177, 'seat_capacity': 60}
-    #     ]
-    # },
-    'COMP590-59': {
-        'title': 'COMP590-59 Class',
-        'sections': [
-            {'section_number': 59, 'seat_capacity': 90}
-        ]
-    },
-    'COMP590-139': {
-        'title': 'COMP590-139 Class',
-        'sections': [
-            {'section_number': 139, 'seat_capacity': 10}
-        ]
-    },
-    'COMP590-140': {
-        'title': 'COMP590-140 Class',
-        'sections': [
-            {'section_number': 140, 'seat_capacity': 280}
-        ]
-    },
-    'COMP590-158': {
-        'title': 'COMP590-158 Class',
-        'sections': [
-            {'section_number': 158, 'seat_capacity': 4}
-        ]
-    },
-    'COMP590-170': {
-        'title': 'COMP590-170 Class',
-        'sections': [
-            {'section_number': 170, 'seat_capacity': 30}
-        ]
-    },
-    'COMP590-172': {
-        'title': 'COMP590-172 Class',
-        'sections': [
-            {'section_number': 172, 'seat_capacity': 30}
-        ]
-    },
-    'COMP590-177': {
-        'title': 'COMP590-177 Class',
-        'sections': [
-            {'section_number': 177, 'seat_capacity': 60}
-        ]
-    },
-    'COMP730': {
-        'title': 'COMP730 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': None}
-        ]
-    },
-    'COMP664': {
-        'title': 'COMP664 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 60}
-        ]
-    },
-    'COMP737': {
-        'title': 'COMP737 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 30}
-        ]
-    },
-    'COMP750': {
-        'title': 'COMP750 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 30}
-        ]
-    },
-    'COMP755': {
-        'title': 'COMP755 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 30}
-        ]
-    },
-    'COMP775': {
-        'title': 'COMP775 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 30}
-        ]
-    },
-    'COMP790-139': {
-        'title': 'COMP790-139 Class',
-        'sections': [
-            {'section_number': 139, 'seat_capacity': 20},
-        ]
-    },
-    'COMP790-144': {
-        'title': 'COMP790-144 Class',
-        'sections': [
-            {'section_number': 144, 'seat_capacity': 30},
-        ]
-    },
-    'COMP790-148': {
-        'title': 'COMP790-148 Class',
-        'sections': [
-            {'section_number': 148, 'seat_capacity': 25},
-        ]
-    },
-    'COMP790-158': {
-        'title': 'COMP790-158 Class',
-        'sections': [
-            {'section_number': 158, 'seat_capacity': 21},
-        ]
-    },
-    'COMP790-170': {
-        'title': 'COMP790-170 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 30},
-        ]
-    },
-    'COMP790-172': {
-        'title': 'COMP790-172 Class',
-        'sections': [
-            {'section_number': 172, 'seat_capacity': 30},
-        ]
-    },
-    'COMP790-173': {
-        'title': 'COMP790-173 Class',
-        'sections': [
-            {'section_number': 173, 'seat_capacity': 30},
-        ]
-    },
-    'COMP790-175': {
-        'title': 'COMP790-175 Class',
-        'sections': [
-            {'section_number': 175, 'seat_capacity': 30},
-        ]
-    },
-    'COMP790-178': {
-        'title': 'COMP790-178 Class',
-        'sections': [
-            {'section_number': 178, 'seat_capacity': 20},
-        ]
-    },
-    'COMP790-183': {
-        'title': 'COMP790-183 Class',
-        'sections': [
-            {'section_number': 183, 'seat_capacity': 30},
-        ]
-    },
-    'COMP790-185': {
-        'title': 'COMP790-185 Class',
-        'sections': [
-            {'section_number': 185, 'seat_capacity': 30},
-        ]
-    },
-    'COMP915': {
-        'title': 'COMP915 Class',
-        'sections': [
-            {'section_number': 1, 'seat_capacity': 20}
-        ]
-    }
-}
-    return courses
-restore_courses()
-
-def load_rooms():
-    rooms = {
-        'SN-0014': {'capacity': 128},
-        'FB-F009': {'capacity': 86},
-        'SN-0011': {'capacity': 66},
-        'FB-F007': {'capacity': 50},
-        'FB-F141': {'capacity': 50},
-        'SN-0115': {'capacity': 25},
-        'FB-F008': {'capacity': 20},
-        'SN-0252': {'capacity': 20},
-        'SN-0006': {'capacity': 15},
-        'SN-0325': {'capacity': 15},
-        'SN-0155': {'capacity': 14},
-        'FB-F120': {'capacity': 12},
-        'SN-0277': {'capacity': 10},
-        'FB-F331': {'capacity': 16},
-        'university': {'capacity': 300}  
-}
-    return rooms
-
-def load_manually_scheduled_classes():
-    manually_scheduled_classes = [
-        # Example of a fully specified manual schedule
-        {
-            'course': 'COMP110',
-            'section': 3,
-            'professor': 'Alyssa Byrnes',
-            'meeting_pattern': 'MWF',
-            'time_slot': ('Monday', 1),
-            'room': 'university'
-        },
-        # Example of a partial specification (only professor assigned)
-        {
-            'course': 'COMP541',
-            'section': 1,
-            'professor': 'Montek Singh',
-            'room': 'FB-F009'
-            # 'meeting_pattern', 'time_slot', and 'room' are unspecified
-        },
-        {'course': 'COMP421',
-        'section': 1,
-        'professor:': "TBD",
+    # Initialize the final structure
+    professors_json = {}
+    for prof in new_profs:
+        professors_json[prof] = {
+            'qualified_courses': [],
+            'availability': [],
+            'max_classes': 0
         }
+
+    # Assign qualified courses to professors
+    for qcourse in qcourses:
+        professor = qcourse['Prof']
+        if professor in professors_json:
+            professors_json[professor]['qualified_courses'].append(qcourse['Course'])
+
+    # Fetch availabilities
+    connection = sqlite3.connect("database.db")
+    connection.row_factory = sqlite3.Row
+    cursor = connection.cursor()
+    try:
+        cursor.execute("SELECT * FROM Availability")
+        availabilities = [dict(row) for row in cursor.fetchall()]
+    except sqlite3.Error as error:
+        print(f"Error occurred while fetching availabilities: {error}")
+        availabilities = []
+    finally:
+        connection.close()
+
+    # Assign availability to professors
+    for availability in availabilities:
+        professor = availability['Prof']
+        if professor in professors_json:
+            professors_json[professor]['availability'].append((availability['AvailableMP'], availability['AvailablePeriod']))
+
+    # Fetch max courses
+    connection = sqlite3.connect("database.db")
+    connection.row_factory = sqlite3.Row
+    cursor = connection.cursor()
+    try:
+        cursor.execute("SELECT * FROM MaxCourses")
+        max_courses = [dict(row) for row in cursor.fetchall()]
+    except sqlite3.Error as error:
+        print(f"Error occurred while fetching max courses: {error}")
+        max_courses = []
+    finally:
+        connection.close()
+
+    # Assign max_classes to professors
+    for m in max_courses:
+        professor = m['Prof']
+        if professor in professors_json:
+            professors_json[professor]['max_classes'] = m['MaxCourses']
+
+    # Return the final dictionary
+    return professors_json
+
+
+import sqlite3
+
+def create_rooms_data():
+    connection = sqlite3.connect("database.db")
+    connection.row_factory = sqlite3.Row
+    
+    cursor = connection.cursor()
+    try:
+        # Fetch both Room and SeatCapacity in a single query
+        query = "SELECT Room, SeatCapacity FROM Rooms"
+        cursor.execute(query)
+        rows = cursor.fetchall()
         
-        # You can add more manually scheduled classes here
-    ]
-    return manually_scheduled_classes
+        # Convert fetched rows into a dictionary
+        rooms_json = {}
+        for row in rows:
+            room_name = row["Room"]
+            capacity = row["SeatCapacity"]
+            rooms_json[room_name] = {
+                "capacity": capacity
+            }
+
+        return rooms_json
+    except sqlite3.Error as error:
+        print(f"Error occurred: {error}")
+        return {}
+    finally:
+        connection.close()
+        print("Database connection closed.")
+
+import sqlite3
+
+def create_courses_data():
+    connection = sqlite3.connect("database.db")
+    connection.row_factory = sqlite3.Row
+
+    cursor = connection.cursor()
+    try:
+        # Fetch all courses and their details
+        query = "SELECT * FROM CoursesAndSections"
+        cursor.execute(query)
+        rows = cursor.fetchall()
+
+        # Build the dictionary
+        courses_json = {}
+        for row in rows:
+            course_name = row["Course"]
+            title = row["Title"]
+            # Convert section to int if it's numeric
+            section_number = int(row["Section"]) if str(row["Section"]).isdigit() else row["Section"]
+            seat_capacity = row["SeatCapacity"]
+
+            # If this is the first time seeing this course, initialize its structure
+            if course_name not in courses_json:
+                courses_json[course_name] = {
+                    "title": title,
+                    "sections": []
+                }
+
+            # Append the section information
+            courses_json[course_name]["sections"].append({
+                "section_number": section_number,
+                "seat_capacity": seat_capacity
+            })
+
+        return courses_json
+    except sqlite3.Error as error:
+        print(f"Error occurred while fetching courses: {error}")
+        return {}
+    finally:
+        connection.close()
+        print("Database connection closed.")
+
+def is_prof_available_for_time_slot(p, mp, period):
+    return (mp, period) in professors[p]['availability']
 
 def is_prof_available_and_qualified(p, c, ts):
-    professors = load_professors()
+    professors = create_professors_data()
     return int(ts in professors[p]['availability'] and c in professors[p]['qualified_courses'])
-
-
-import pulp
 
 def run_scheduling_algorithm (
     days,
@@ -934,9 +238,8 @@ def run_scheduling_algorithm (
     time_slots_mp = {}
     for mp in meeting_patterns:
         mp_time_slots = []
-        for day in meeting_patterns[mp]['days']:
-            for period in meeting_patterns[mp]['periods'].keys():
-                mp_time_slots.append((day, period))
+        for period in meeting_patterns[mp]['periods'].keys():
+            mp_time_slots.append((mp, period))
         time_slots_mp[mp] = mp_time_slots
 
     # Identify small and large classes based on seat capacity
@@ -954,12 +257,14 @@ def run_scheduling_algorithm (
 
     # Generate all valid combinations of indices for x
     x_indices = []
+    
     for c in courses:
         for section in courses[c]['sections']:
             s = section['section_number']
             seat_capacity = section['seat_capacity']
             for mp in possible_meeting_patterns:
-                for ts in time_slots_mp[mp]:
+                for period in meeting_patterns[mp]['periods'].keys():
+                    ts = (mp, period)
                     for r in rooms.keys():
                         # Enforce large classes to use 'university' room only
                         if (c, s) in large_classes and r != 'university':
@@ -970,7 +275,7 @@ def run_scheduling_algorithm (
                         # Enforce room capacity constraints
                         if seat_capacity is not None and rooms[r]['capacity'] < seat_capacity:
                             continue  # Skip this combination
-                        idx = (c, s, mp, ts, r)
+                        idx = (c, s, mp, period, r)
                         x_indices.append(idx)
 
     # Define the decision variables
@@ -1008,59 +313,44 @@ def run_scheduling_algorithm (
             p = entry['professor']
             # Ensure the professor is assigned to the class
             prob += z[(c, s, p)] == 1, f"Manual_Professor_Assigned_{c}_{s}_{p}"
-            # Ensure the professor is not assigned to other classes at the same time
-            if 'time_slot' in entry:
-                ts = tuple(entry['time_slot'])
-                day, period = ts
-                # Professors can't teach other classes at the same time
-                for other_c in courses:
-                    for section in courses[other_c]['sections']:
-                        other_s = section['section_number']
-                        if (other_c, other_s, p) in z and (other_c != c or other_s != s):
-                            # Ensure uniqueness of constraint names
-                            constraint_name = f"Manual_Prof_Time_Conflict_{p}_{day}_{period}_{other_c}_{other_s}"
-                            prob += pulp.lpSum([
-                                x[idx]
-                                for idx in x_indices
-                                if idx[0] == other_c and idx[1] == other_s and idx[3] == ts
-                            ]) == 0, constraint_name
 
-        # Fix meeting pattern, time slot, and room if specified
-        if 'meeting_pattern' in entry and 'time_slot' in entry and 'room' in entry:
+        # Fix meeting pattern, period, and room if specified
+        if 'meeting_pattern' in entry and 'period' in entry and 'room' in entry:
             mp = entry['meeting_pattern']
-            ts = tuple(entry['time_slot'])
+            period = entry['period']
             r = entry['room']
-            # Set x variable corresponding to these values to 1
-            idx = (c, s, mp, ts, r)
+            idx = (c, s, mp, period, r)
             if idx in x:
-                prob += x[idx] == 1, f"Manual_Schedule_{c}_{s}_{mp}_{ts}_{r}"
+                prob += x[idx] == 1, f"Manual_Schedule_{c}_{s}_{mp}_{period}_{r}"
             else:
-                pass  
+                print(f"Warning: Invalid manual schedule for class {c} section {s}.")
         else:
-            # Partial specifications
+            # Fix meeting pattern if specified
             if 'meeting_pattern' in entry:
                 mp = entry['meeting_pattern']
                 prob += pulp.lpSum([
                     x[idx]
                     for idx in x_indices
                     if idx[0] == c and idx[1] == s and idx[2] == mp
-                ]) == 1, f"Manual_Meeting_Pattern_{c}_{s}_{mp}"
+                ]) == y[(c, s)], f"Manual_Meeting_Pattern_{c}_{s}_{mp}"
 
-            if 'time_slot' in entry:
-                ts = tuple(entry['time_slot'])
+            # Fix period if specified
+            if 'period' in entry:
+                period = entry['period']
                 prob += pulp.lpSum([
                     x[idx]
                     for idx in x_indices
-                    if idx[0] == c and idx[1] == s and idx[3] == ts
-                ]) == 1, f"Manual_Time_Slot_{c}_{s}_{ts}"
+                    if idx[0] == c and idx[1] == s and idx[3] == period
+                ]) == y[(c, s)], f"Manual_Period_{c}_{s}_{period}"
 
+            # Fix room if specified
             if 'room' in entry:
                 r = entry['room']
                 prob += pulp.lpSum([
                     x[idx]
                     for idx in x_indices
                     if idx[0] == c and idx[1] == s and idx[4] == r
-                ]) == 1, f"Manual_Room_{c}_{s}_{r}"
+                ]) == y[(c, s)], f"Manual_Room_{c}_{s}_{r}"
 
     # -----------------------------
     # Constraints
@@ -1098,15 +388,17 @@ def run_scheduling_algorithm (
 
     # 4. A professor cannot teach more than one class at the same time
     for p in professors:
-        for ts in set(professors[p]['availability']):
-            day, period = ts
-            ts_str = f"{day}_{period}"
-            constraint_name = f"Prof_Time_Conflict_{p}_{ts_str}"
-            prob += pulp.lpSum([
-                x[idx]
-                for idx in x_indices
-                if idx[3] == ts and (idx[0], idx[1], p) in z
-            ]) <= 1, constraint_name
+        for mp in meeting_patterns:
+            for period in meeting_patterns[mp]['periods']:
+                constraint_name = f"Prof_Time_Conflict_{p}_{mp}_{period}"
+                # Sum over all classes assigned to professor p at (mp, period)
+                prob += pulp.lpSum([
+                    x[idx]
+                    for idx in x_indices
+                    if (idx[0], idx[1], p) in z
+                    and idx[2] == mp
+                    and idx[3] == period
+                ]) <= 1, constraint_name
 
     # 5a. If z[(c, s, p)] == 1, then x[idx] == 1 for some idx
     for c in courses:
@@ -1115,46 +407,47 @@ def run_scheduling_algorithm (
             for p in professors:
                 if (c, s, p) in z:
                     constraint_name = f"Link_z_x_{c}_{s}_{p}"
+                    
                     prob += pulp.lpSum([
                         x[idx]
                         for idx in x_indices
-                        if idx[0] == c and idx[1] == s and idx[3] in professors[p]['availability']
+                        if idx[0] == c and idx[1] == s and is_prof_available_for_time_slot(p, idx[2], idx[3])
                     ]) >= z[(c, s, p)], constraint_name
 
     # 5b. If x[idx] == 1, then z[(c, s, p)] == 1 for some p
     for idx in x_indices:
-        c, s, mp, ts, r = idx
-        constraint_name = f"Link_x_z_{c}_{s}_{mp}_{ts}_{r}"
+        c, s, mp, period, r = idx
+        constraint_name = f"Link_x_z_{c}_{s}_{mp}_{period}_{r}"
+        
         prob += x[idx] <= pulp.lpSum([
             z[(c, s, p)]
             for p in professors
-            if (c, s, p) in z and ts in professors[p]['availability']
+            if (c, s, p) in z and is_prof_available_for_time_slot(p, mp, period)
         ]), constraint_name
 
     # 6. Room capacity constraints (no double booking)
-    # Collect all unique time slots across all meeting patterns
-    all_time_slots = set()
-    for mp in time_slots_mp:
-        all_time_slots.update(time_slots_mp[mp])
+    # Collect all unique day-period combinations
+    all_day_periods = set()
+    for mp in meeting_patterns:
+        days = meeting_patterns[mp]['days']
+        periods = meeting_patterns[mp]['periods'].keys()
+        for day in days:
+            for period in periods:
+                all_day_periods.add((day, period))
 
+    # 6. Room capacity constraints (no double booking)
     for r in rooms:
         if r != 'university':
-            for ts in all_time_slots:
-                day, period = ts
-                ts_str = f"{day}_{period}"
-                constraint_name = f"Room_Capacity_{r}_{ts_str}"
-                # Ensure uniqueness of constraint names
-                if constraint_name in prob.constraints:
-                    index = 1
-                    while f"{constraint_name}_{index}" in prob.constraints:
-                        index += 1
-                    constraint_name = f"{constraint_name}_{index}"
-                prob += pulp.lpSum([
-                    x[idx]
-                    for idx in x_indices
-                    if idx[4] == r and idx[3] == ts
-                ]) <= 1, constraint_name
-
+            for mp in meeting_patterns:
+                for period in meeting_patterns[mp]['periods']:
+                    constraint_name = f"Room_Capacity_{r}_{mp}_{period}"
+                    prob += pulp.lpSum([
+                        x[idx]
+                        for idx in x_indices
+                        if idx[4] == r
+                        and idx[2] == mp
+                        and idx[3] == period
+                    ]) <= 1, constraint_name
     # 7. Room overlap constraints
     for r in rooms:
         for d in days:
@@ -1251,7 +544,7 @@ def run_scheduling_algorithm (
                     assigned = False
                     for idx in x_indices:
                         if idx[0] == c and idx[1] == s and pulp.value(x[idx]) == 1:
-                            c, s, mp, ts, r = idx
+                            c, s, mp, period, r = idx
                             section = next(sec for sec in courses[c]['sections'] if sec['section_number'] == s)
                             seat_capacity = section['seat_capacity']
                             # Find the assigned professor
@@ -1260,19 +553,17 @@ def run_scheduling_algorithm (
                                 if (c, s, p) in z and pulp.value(z[(c, s, p)]) == 1:
                                     assigned_professor = p
                                     break
-                            day, period = ts
+                            days = meeting_patterns[mp]['days']
                             start_time = meeting_patterns[mp]['periods'][period]['start_time']
-                            end_time = meeting_patterns[mp]['periods'][period]['end_time']
                             schedule.append({
                                 'Course': c,
                                 'Section': s,
                                 'Title': courses[c]['title'],
                                 'Professor': assigned_professor,
                                 'Meeting Pattern': mp,
-                                'Day': day,
+                                'Days': days,
                                 'Period': period,
                                 'Start Time': start_time,
-                                'End Time': end_time,
                                 'Room': r,
                                 'Seat Capacity': seat_capacity
                             })
@@ -1280,18 +571,10 @@ def run_scheduling_algorithm (
                             assigned = True
                             break
                     if not assigned:
-                        unscheduled_classes.append({
-                            'Course': c,
-                            'Section': s,
-                            'Title': courses[c]['title']
-                        })
+                        unscheduled_classes.append((c, s))
                 else:
                     # Class is not scheduled
-                    unscheduled_classes.append({
-                        'Course': c,
-                        'Section': s,
-                        'Title': courses[c]['title']
-                    })
+                    unscheduled_classes.append((c, s))
 
         # Initialize a dictionary to count classes assigned to each professor
         professor_class_counts = {p: 0 for p in professors}
@@ -1326,22 +609,21 @@ def run_scheduling_algorithm (
             'message': f"No feasible solution found. Solver Status: {pulp.LpStatus[prob.status]}"
         }
         return result
-'''
+
+
 manually_scheduled_classes = None
 days = load_days()
 mwf_periods, tth_periods, mw_periods = load_periods()
 meeting_patterns =  load_meeting_patterns(mwf_periods, tth_periods, mw_periods)
-professors =  load_professors()
-courses =  load_courses()
-rooms =  load_rooms()
+professors =  create_professors_data()
+courses =  create_courses_data()
+rooms =  create_rooms_data()
 
 results = run_scheduling_algorithm(
-    days=days,
-    meeting_patterns=meeting_patterns,
-    professors=professors,
-    courses=courses,
-    rooms=rooms,
-    manually_scheduled_classes=manually_scheduled_classes
+days=days,
+meeting_patterns=meeting_patterns,
+professors=professors,
+courses=courses,
+rooms=rooms,
+manually_scheduled_classes=manually_scheduled_classes
 )
-print(results)
-'''
